trying to fix GetAcquiredData

+++ Andor/camera.pyx
@@ -131,8 +131,8 @@ class Camera:
 
     def _GetAcquiredData(self,width,height):
         cdef unsigned int size = width*height
-        cdef np.ndarray[np.int, ndim=1, mode = 'c'] data = np.ascontiguousarray(np.zeros(size), dtype = np.int)
-        cdef int* data_ptr = <int*> data.data
+        cdef np.ndarray[np.int, ndim=1, mode = 'c'] data = np.ascontiguousarray(np.zeros(size), dtype = np.long)
+        cdef long* data_ptr = <long*> data.data
         #cdef np.ndarray[np.int,mode="c"] data = np.zeros(size,dtype=np.int)
         #cdef int* data_ptr = &data[0]
         print(size)
